针对cython进行优化

+++ setup.py
@@ -9,16 +9,19 @@ import shutil
 
 #是否编译Source中的游戏本体
 compile_py_files_in_Source = False
+debug_c = False
 
 #生成Zero引擎的c和pyd文件
 for path in glob.glob(r'Zero3/*.pyx'):
     setup(ext_modules=cythonize(path,show_all_warnings=True))
     #删除Zero引擎的c文件
-    os.remove(path.replace(".pyx",".c"))
+    if not debug_c:
+        os.remove(path.replace(".pyx",".c"))
 #生成游戏本体源代码的c和pyd文件
 if compile_py_files_in_Source:
     for path in glob.glob(r'Source/*.py'):
         setup(ext_modules=cythonize(path))
-        os.remove(path.replace(".pyx",".c"))
+        if not debug_c:
+            os.remove(path.replace(".pyx",".c"))
 #删除build文件夹
 shutil.rmtree('build')
